Make dbus connection mandatory
