+++ rust/agama-utils/Cargo.toml
@@ -5,6 +5,7 @@ rust-version.workspace = true
 edition.workspace = true
 
 [dependencies]
+agama-locale-data = { path = "../agama-locale-data" }
 async-trait = "0.1.89"
 serde = { version = "1.0.228", features = ["derive"] }
 serde_json = { version = "1.0.140", features = ["raw_value"] }
@@ -17,7 +18,7 @@ utoipa = "5.3.1"
 zbus = "5.7.1"
 zvariant = "5.5.2"
 gettext-rs = { version = "0.7.2", features = ["gettext-system"] }
-agama-locale-data = { path = "../agama-locale-data" }
+uuid = { version = "1.10.0", features = ["v4"] }
 
 [dev-dependencies]
 tokio-test = "0.4.4"
