+++ rust/agama-utils/src/issue/start.rs
@@ -18,10 +18,14 @@
 // To contact SUSE LLC about this file by physical or electronic mail, you may
 // find current contact information at www.suse.com.
 
-use crate::actor::{self, Handler};
-use crate::api::event;
-use crate::issue::monitor::{self, Monitor};
-use crate::issue::service::{self, Service};
+use crate::{
+    actor::{self, Handler},
+    api::event,
+    issue::{
+        monitor::{self, Monitor},
+        service::{self, Service},
+    },
+};
 
 #[derive(thiserror::Error, Debug)]
 pub enum Error {
@@ -31,26 +35,28 @@ pub enum Error {
 
 pub async fn start(
     events: event::Sender,
-    dbus: Option<zbus::Connection>,
+    dbus: zbus::Connection,
 ) -> Result<Handler<Service>, Error> {
     let service = Service::new(events);
     let handler = actor::spawn(service);
 
-    if let Some(conn) = dbus {
-        let dbus_monitor = Monitor::new(handler.clone(), conn);
-        monitor::spawn(dbus_monitor);
-    }
+    let dbus_monitor = Monitor::new(handler.clone(), dbus);
+    monitor::spawn(dbus_monitor);
 
     Ok(handler)
 }
 
 #[cfg(test)]
 mod tests {
-    use crate::api::event::Event;
-    use crate::api::issue::{Issue, IssueSeverity, IssueSource};
-    use crate::api::scope::Scope;
-    use crate::issue;
-    use crate::issue::message;
+    use crate::{
+        api::{
+            event::Event,
+            issue::{Issue, IssueSeverity, IssueSource},
+            scope::Scope,
+        },
+        issue::{self, message},
+        test,
+    };
     use tokio::sync::broadcast::{self, error::TryRecvError};
 
     fn build_issue() -> Issue {
@@ -66,7 +72,8 @@ mod tests {
     #[tokio::test]
     async fn test_get_and_update_issues() -> Result<(), Box<dyn std::error::Error>> {
         let (events_tx, mut events_rx) = broadcast::channel::<Event>(16);
-        let issues = issue::start(events_tx, None).await.unwrap();
+        let dbus = test::dbus::connection().await.unwrap();
+        let issues = issue::start(events_tx, dbus).await.unwrap();
         let issues_list = issues.call(message::Get).await.unwrap();
         assert!(issues_list.is_empty());
 
@@ -85,7 +92,8 @@ mod tests {
     #[tokio::test]
     async fn test_update_without_event() -> Result<(), Box<dyn std::error::Error>> {
         let (events_tx, mut events_rx) = broadcast::channel::<Event>(16);
-        let issues = issue::start(events_tx, None).await.unwrap();
+        let dbus = test::dbus::connection().await.unwrap();
+        let issues = issue::start(events_tx, dbus).await.unwrap();
 
         let issues_list = issues.call(message::Get).await.unwrap();
         assert!(issues_list.is_empty());
@@ -104,7 +112,8 @@ mod tests {
     #[tokio::test]
     async fn test_update_without_change() -> Result<(), Box<dyn std::error::Error>> {
         let (events_tx, mut events_rx) = broadcast::channel::<Event>(16);
-        let issues = issue::start(events_tx, None).await.unwrap();
+        let dbus = test::dbus::connection().await.unwrap();
+        let issues = issue::start(events_tx, dbus).await.unwrap();
 
         let issue = build_issue();
         let update = message::Update::new(Scope::Manager, vec![issue.clone()]);
