+++ rust/agama-utils/src/lib.rs
@@ -31,3 +31,4 @@ pub mod issue;
 pub mod openapi;
 pub mod progress;
 pub mod question;
+pub mod test;
